LPS-79599 move to the correct package
