+++ modules/core/portal-compat/blogs/portal-blogs-compat-test/src/testIntegration/java/com/liferay/portal/blogs/compat/attachments/test/BlogsEntryAttachmentFileEntryHelperTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.blogs.attachments.test;
+package com.liferay.portal.blogs.compat.attachments.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.blogs.model.BlogsEntry;
