+++ modules/core/portal-compat/blogs/portal-blogs-compat-test/src/testIntegration/java/com/liferay/portal/blogs/compat/search/test/BlogsEntrySearchTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.blogs.search.test;
+package com.liferay.portal.blogs.compat.search.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.blogs.model.BlogsEntry;
