+++ modules/core/portal-compat/blogs/portal-blogs-compat-test/src/testIntegration/java/com/liferay/portal/blogs/compat/subscription/test/BlogsSubscriptionRootContainerModelTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.blogs.subscription.test;
+package com.liferay.portal.blogs.compat.subscription.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.blogs.model.BlogsEntry;
@@ -25,24 +25,51 @@ import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
 import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.SynchronousMailTestRule;
-import com.liferay.portlet.subscriptions.test.BaseSubscriptionAuthorTestCase;
+import com.liferay.portlet.subscriptions.test.BaseSubscriptionRootContainerModelTestCase;
 
 import org.junit.ClassRule;
+import org.junit.Ignore;
 import org.junit.Rule;
+import org.junit.Test;
 import org.junit.runner.RunWith;
 
 /**
- * @author José Ángel Jiménez
+ * @author Sergio González
+ * @author Roberto Díaz
  */
 @RunWith(Arquillian.class)
-public class BlogsSubscriptionAuthorTest
-	extends BaseSubscriptionAuthorTestCase {
+public class BlogsSubscriptionRootContainerModelTest
+	extends BaseSubscriptionRootContainerModelTestCase {
 
 	@ClassRule
 	@Rule
 	public static final AggregateTestRule aggregateTestRule =
 		new AggregateTestRule(
 			new LiferayIntegrationTestRule(), SynchronousMailTestRule.INSTANCE);
+
+	@Ignore
+	@Override
+	@Test
+	public void testSubscriptionRootContainerModelWhenAddingBaseModelInContainerModel() {
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testSubscriptionRootContainerModelWhenAddingBaseModelInSubcontainerModel() {
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testSubscriptionRootContainerModelWhenUpdatingBaseModelInContainerModel() {
+	}
+
+	@Ignore
+	@Override
+	@Test
+	public void testSubscriptionRootContainerModelWhenUpdatingBaseModelInSubcontainerModel() {
+	}
 
 	@Override
 	protected long addBaseModel(long userId, long containerModelId)
@@ -63,10 +90,11 @@ public class BlogsSubscriptionAuthorTest
 	}
 
 	@Override
-	protected void addSubscription(long userId, long containerModelId)
+	protected void addSubscriptionContainerModel(long containerModelId)
 		throws Exception {
 
-		BlogsEntryLocalServiceUtil.subscribe(userId, group.getGroupId());
+		BlogsEntryLocalServiceUtil.subscribe(
+			user.getUserId(), group.getGroupId());
 	}
 
 	@Override
